ai_modules/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
psycopg2-binary>=2.9
redis>=5.0
minio>=7.2
numpy>=1.26
opencv-python-headless>=4.9
scikit-image>=0.22
Pillow>=10.2
ultralytics>=8.3
shapely>=2.0
python-dotenv>=1.0
requests>=2.31
